refactor: move tag adder to root component dir
